requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.5
python-dotenv>=1.0.1
gspread>=6.0.2
google-auth>=2.33.0
